+++ app/schemas/models.py
@@ -3,12 +3,12 @@ from typing import List, Literal, Optional
 from openai.types import Model
 from pydantic import BaseModel
 
-from app.utils.variables import EMBEDDINGS_MODEL_TYPE, LANGUAGE_MODEL_TYPE
+from app.utils.variables import EMBEDDINGS_MODEL_TYPE, LANGUAGE_MODEL_TYPE, AUDIO_MODEL_TYPE
 
 
 class Model(Model):
     max_context_length: Optional[int] = None
-    type: Literal[LANGUAGE_MODEL_TYPE, EMBEDDINGS_MODEL_TYPE]
+    type: Literal[LANGUAGE_MODEL_TYPE, EMBEDDINGS_MODEL_TYPE, AUDIO_MODEL_TYPE]
     status: Literal["available", "unavailable"]